% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{bread.kclass}
\alias{bread.kclass}
\title{S3 method for extracting the 'bread' portion of a sandwich estimator.}
\usage{
bread.kclass(object, ...)
}
\arguments{
\item{object}{the fitted kclass model object.}

\item{...}{additional parameters.}
}
\value{
A matrix containing an estimator for the expectation of the negative
  derivative of the estimating functions, usually a Hessian matrix.
}
\description{
S3 method for extracting the 'bread' portion of a sandwich estimator.
}
\seealso{
sandwich
}

