% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{estfun.kclass}
\alias{estfun.kclass}
\title{S3 method for a kclass estimator function}
\usage{
estfun.kclass(object, ...)
}
\arguments{
\item{object}{the fitted kclass model object.}

\item{...}{other parameters passed to methods.}
}
\value{
A matrix containing the empirical estimating functions.
}
\description{
S3 method for a kclass estimator function
}
\seealso{
estfun
}

