% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{hatvalues.kclass}
\alias{hatvalues.kclass}
\title{S3 method to get the hat values for a model.}
\usage{
\method{hatvalues}{kclass}(object, full = FALSE)
}
\arguments{
\item{object}{a fitted model of type kclass.}

\item{full}{logical to return the full matrix or just the diagonal.}
}
\value{
a diagonal (or matrix) of estimated values useful in alternative
}
\description{
S3 method to get the hat values for a model.
}

