% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kclass.R
\name{kclass.fit}
\alias{kclass.fit}
\title{Low-level kclass instrumental variables function.}
\usage{
kclass.fit(x, y, z = NULL, weights = NULL, offset = NULL, k = NULL,
  alpha = NULL, model.type = c("TSLS", "LIML", "FULLER", "KCLASS", "OLS"),
  eig = c("eigen", "geigen"), na.action, ...)
}
\arguments{
\item{x}{data frame or matrix of endogenous regressors.}

\item{y}{data frame or vector of the dependent variable.}

\item{z}{data frame or matrix of instruments and/or exogenous regressors.}

\item{weights}{sample weights. Default is NULL.}

\item{offset}{model offset.}

\item{k}{the k parameter value. If k is set, \code{model.type} is set to "KCLASS".}

\item{alpha}{the fuller parameter, only used if \code{model.type} is "FULLER".}

\item{model.type}{string determining model type. Must be one of "OLS", "LIML",
"TSLS", "KCLASS", "FULLER". Default is LIML."}

\item{eig}{eigenvalue method. One of "eigen" or "geigen".}

\item{na.action}{function indicating what to do when data contain NAs. The
default is set by the na.action setting of options}

\item{...}{additional parameters not currently used.}
}
\value{
a model object of class kclass.
}
\description{
Low-level kclass instrumental variables function.
}

