% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{model.matrix.kclass}
\alias{model.matrix.kclass}
\title{S3 method to extract model matrix from kclass objects.}
\usage{
\method{model.matrix}{kclass}(object, component = c("projected", "regressors",
  "instruments"))
}
\arguments{
\item{object}{a model of class kclass.}

\item{component}{indicator for regressors, instruments, or projected values.}
}
\value{
the model matrix of the object.
}
\description{
S3 method to extract model matrix from kclass objects.
}

