% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{predict.kclass}
\alias{predict.kclass}
\title{Predict values based on a fitted kclass model}
\usage{
\method{predict}{kclass}(object, newdata, na.action = na.pass, ...)
}
\arguments{
\item{object}{a model object of class 'kclass'.}

\item{newdata}{a data frame or vector. If no data is submitted, the original
fitted values are returned.}
}
\value{
a vector of predicted values.
}
\description{
Generates a vector of outcome predictions based on new regressor data. The
new data must be a data frame or matrix with the same number of columns as
coefficients in the model.
}

