% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.kclass}
\alias{print.kclass}
\title{S3 print method for class 'kclass'}
\usage{
\method{print}{kclass}(object, digits = 3, print.gap = 2, quote = FALSE,
  ...)
}
\arguments{
\item{object}{the model object returned by \code{kclass()}.}

\item{digits}{number of digits to display.}

\item{print.gap}{the gap between each printed coefficient.}

\item{quote}{logical indicating whether to quote the coefficient values.}

\item{...}{other parameters passed to \code{print.default}.}
}
\value{
the object.
}
\description{
Formats the object and prints some information when the object is entered
by itself in the console.
}

