% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.summary.kclass}
\alias{print.summary.kclass}
\title{S3 print method for class 'summary.kclass'}
\usage{
\method{print}{summary.kclass}(object, digits = 3,
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{an object with class 'summary.kclass', usually ouput from \code{summary()}.}

\item{digits}{the number of digits to display.}

\item{signif.stars}{logical to show significance indicators.}

\item{...}{other parameters passed to \code{\link{printCoefmat}}.}
}
\value{
the summary object.
}
\description{
S3 print method for class 'summary.kclass'
}

