% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{qob}
\alias{qob}
\title{Weekly earnings, education, and age.}
\format{a data frame with 329508 rows and 6 columns:
\tabular{ll}{
 \code{LWKLYWGE} \tab Log weekly wage in dollars. \cr
 \code{EDUC} \tab Years of education. \cr
 \code{AGE} \tab Age in years. \cr
 \code{AGESQ} \tab Age squared. \cr
 \code{YOB} \tab The last two digist of year of birth. \cr
 \code{QOB} \tab Quarter of birth. \cr
}}
\source{
\url{http://economics.mit.edu/faculty/angrist/data1/data/angkru1991}
}
\usage{
qob
}
\description{
A dataset containing 1980 census data used in Angrist and Krueger 1991.
The data include Weekly earnings, education, age, year of birth and
quarter of birth.
}
\keyword{datasets}

