% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rkclass.R
\docType{package}
\name{rivreg}
\alias{rivreg}
\alias{rivreg-package}
\title{rkclass-package: Perform instrumental variables estimates.}
\description{
This package contains functions, primarily kclass, for instrumental
 variables regression. In particular, it focuses on corrected standard errors
 (CSE) under many instruments, as described in Hansen, Hausman, and Newey 2004.
 It also installs an example dataset for testing from Angrist and Krueger 1991 (~3MB).
}
\author{
Nicholas Potter
}

