% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.kclass}
\alias{summary.kclass}
\title{S3 summary method for 'kclass' objects}
\usage{
\method{summary}{kclass}(object, vcov. = NULL, df = NULL,
  errors = c("standard", "const", "robust", "bw", "white", "HAC", "CSE", "HC",
  "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5"), ...)
}
\arguments{
\item{object}{a model object of class 'kclass'}

\item{vcov.}{a function or matrix.}

\item{df}{the degrees of freedom.}

\item{errors}{convenience method for calculating adjusted standard errors.
Use \code{robust} for heterskedastic-robust errors, \code{bw} for autocorrelation-
robust errors, \code{white} for White's estimator. \code{bekker} provides Bekker's
adjusted standard errors (see Bekker 1994), while CSE provides the
corrected standard errors of Hansen, Hausman, and Newey 2004. Other
allowed options include any of the \code{type} options available in
\code{\link{sandwich}{vcovHC}}.}

\item{...}{other parameters passed to print function.}

\item{robust}{logical to report robust sandwich estimator standard errors.
These are heteroskedastic robust.}

\item{bw}{logical to report XX standard errors. These are autocorrelation
robust. Setting both robust and bw to \code{TRUE} reports standard errors
that are both heteroskedastic and autocorrelation robust.}
}
\value{
a list object containing the following values:
\tabular{ll}{
  \code{value} \tab explanation
}
}
\description{
S3 summary method for 'kclass' objects
}

