% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{terms.kclass}
\alias{terms.kclass}
\title{S3 terms method for class 'kclass'}
\usage{
\method{terms}{kclass}(object, component = c("regressors", "instruments"),
  ...)
}
\arguments{
\item{object}{a model object of class kclass.}

\item{component}{either "regressors" or "instruments".}

\item{...}{other parameters not yet used.}
}
\value{
model terms for component.
}
\description{
S3 terms method for class 'kclass'
}

