% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{vcov.kclass}
\alias{vcov.kclass}
\title{S3 vcov method for class 'kclass'}
\usage{
\method{vcov}{kclass}(object, vcov. = NULL, errors = NULL)
}
\arguments{
\item{object}{a model object of class 'kclass'.}
}
\value{
a k * k matrix of variances and covariances, where k is the
  number of coefficients in the model.
}
\description{
Calculates the variance-covariance matrix for a kclass model.
}

