% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov_bekker}
\alias{vcov_bekker}
\title{Calculate the Bekker-adjusted variance-covariance matrix.}
\usage{
vcov_bekker(object, ...)
}
\arguments{
\item{object}{a fitted model of class kclass or ivreg.}
}
\value{
a variance-covariance matrix using Bekker's estimator.
  See Bekker 1994 for more details.
}
\description{
Calculate the Bekker-adjusted variance-covariance matrix.
}
\references{
\cite{HansenHausmanNewye2004}
}

